BUG: Rename vendored ujson module from json to _ujson
The vendored ujson C extension was registered as `pandas._libs.json`,
which collides with the stdlib `json` module. This caused segfaults
and import-order-dependent bugs when other C extensions triggered
numpy initialization during module loading.

Rename to `pandas._libs._ujson` to avoid the collision. The leading
underscore signals it is a private C extension.
